create useCreateEditMember hook, add CreateEditMemberInputs
