+++ packages/atlas/src/components/_auth/CreateEditMemberInputs/CreateEditMemberInputs.styles.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled'
+
+import { TextField } from '@/components/_inputs/TextField'
+import { sizes } from '@/styles'
+
+export const StyledTextField = styled(TextField)`
+  margin-bottom: ${sizes(5)};
+`
