+++ packages/atlas/src/components/_auth/CreateEditMemberInputs/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateEditMemberInputs'
